requirements.txt
pandas 
numpy 
scikit-learn
tensorboard
matplotlib
streamlit
scikeras
nltk
dash
tensorflow-cpu==2.13.0
keras==2.13.1
protobuf==3.20.*
numpy
pickle-mixin
gunicorn
